+++ demo/component/Common.js
@@ -2,10 +2,12 @@ import Demobox from './Demobox';
 import ExampleHeader from './ExampleHeader';
 import ObjectView from './ObjectView';
 import PropsTable from './PropsTable';
+import Method from './Method';
 
 export {
   Demobox,
   ExampleHeader,
   ObjectView,
-  PropsTable
+  PropsTable,
+  Method
 };