+++ demo/routes/Breadcrumb/index.js
@@ -1,5 +1,6 @@
 export default {
   path: 'breadcrumb',
+  breadcrumbName: '面包屑',
   getComponent(nextState, cb) {
     require.ensure([], (require) => {
       cb(null, require('./components/Breadcrumb').default);