+++ demo/routes/Clipboard/components/Clipboard.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Flex, Card } from 'react-timo';
-import {ExampleHeader, PropsTable, Demobox} from 'demo-component';
+import {ExampleHeader, PropsTable, Demobox, Method} from 'demo-component';
 
 import clipboardProps from '!!docgen-loader!../../../../src/component/Clipboard';
 import basicCode from '!!raw-loader!./Basic';
@@ -16,7 +16,16 @@ class ClipboardExample extends Component {
       <Flex column className="demo-clipboard" flex={1}>
         <ExampleHeader
           title="Clipboard Component"
-          desc="Simple copy the string to your clipboard"
+          desc={
+            <div>
+              Simple copy the string to your clipboard base on&nbsp;
+              <a
+                target="_blank"
+                src="https://github.com/sudodoki/copy-to-clipboard"
+                style={{color: '#03A9F4'}}
+              >copy-to-clipboard</a>
+            </div>
+          }
         />
         <Card title="Basic Example">
           <Demobox code={basicCode}>
@@ -24,6 +33,7 @@ class ClipboardExample extends Component {
           </Demobox>
         </Card>
         <PropsTable list={clipboardProps}/>
+        <Method list={clipboardProps} />
       </Flex>
     );
   }