+++ src/component/Clipboard/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import copy from 'copy-to-clipboard';
+import copyPackage from 'copy-to-clipboard';
 import _isFunction from 'lodash/isFunction';
 import _noop from "lodash/noop";
 
@@ -32,6 +32,14 @@ class Clipboard extends React.PureComponent {
     onCopy: _noop
   };
 
+  /**
+   * copy the text
+   * @param {string} str      the text need to copy
+   * @param {object} options  copy-to-clipboard options
+   */
+  static copy(str, options) {
+    copyPackage(str, options);
+  }
 
   onClick = event => {
     const {
@@ -43,7 +51,7 @@ class Clipboard extends React.PureComponent {
 
     const elem = React.Children.only(children);
 
-    const result = copy(text, options);
+    const result = Clipboard.copy(text, options);
 
     if (onCopy) {
       onCopy(text, result);